+++ src/main/java/ch/fhnw/deardevbackend/mapper/DashboardMapper.java
@@ -9,7 +9,7 @@ import org.mapstruct.factory.Mappers;
 public interface DashboardMapper {
     DashboardMapper INSTANCE = Mappers.getMapper(DashboardMapper.class);
 
-    default DashboardDTO toDashboardDTO(Integer workKindId, String workKindName, Long voteCount,  Integer averageScore, Integer happinessScore) {
+    default DashboardDTO toDashboardDTO(Integer workKindId, String workKindName, Integer voteCount,  Integer averageScore, Integer happinessScore) {
         DashboardDTO.WorkKindDTO workKindDTO = new DashboardDTO.WorkKindDTO(workKindId, workKindName, voteCount, happinessScore);
 
         return new DashboardDTO(workKindDTO, averageScore);
